+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## UNRELEASED
 ### Added
 * ui: Add listener to `SelfScanningFragment` to react to camera permission changes
+* core: Receive origin type from `PaymentMethodDesricptor` in case it is not set by default
 ### Changed
 ### Removed
 ### Fixed